requirements.txt
pandas
numpy
matplotlib
plotly
streamlit-aggrid
pulp